Add files via upload
This is a copy of my Master's thesis. The objective of the thesis was to estimate the emotional content of an image based on the observer's eye scan pattern. A classification model was built using a machine learning algorithm using MATLAB. The input imageset was trained on a Convolutional Neural Network to improve the efficiency.